Pass correct Tickets URL to Nav, NavBlock
In the same style as the previous commit, but this time for client-side navigation. Being able to pass a Sanity-fetched URL required moving the header and Nav into [[...slug]].tsx.
